% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_vessel.R
\name{display_vessel}
\alias{display_vessel}
\title{Display information associated to a vessel ID in pretty tables}
\usage{
display_vessel(id = "FRA000669307", type = "cfr")
}
\arguments{
\item{id}{A vessel identifier (CFR, MMSI, UVI, IRCS, external marking or registration number)}

\item{type}{identifier type, either "cfr", "mmsi", "uvi", "ircs", "external_marking" or "registration_number"}
}
\value{
The function prints tables, one per CFR associated to a vessel ID
}
\description{
Display information associated to a vessel ID in pretty tables
}
