% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_events.R
\name{get_all_events}
\alias{get_all_events}
\title{Get all events records in BLOOM fleet register}
\usage{
get_all_events()
}
\value{
The function returns a dataframe containing all the events saved by BLOOM
}
\description{
Get all events records in BLOOM fleet register
}
