% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Get records in BLOOM fleet register}
\usage{
get_events(
  country_iso3 = NULL,
  start_date = "1980-01-01",
  end_date = Sys.Date()
)
}
\arguments{
\item{country_iso3}{A string ISO3 code only for European country. One of "BEL", "BGR", "HRV", "CYP", "DNK", "EST", "FIN", "FRA", DEU", "GRC", "GBR", "IRL", "ITA", "LVA", "LTU", "MLT", "NLD", "POL", "PRT", "ROU", "SVN", "ESP", "SWE"}

\item{start_date}{A string date in the following format: \%Y-\%m-\%d}

\item{end_date}{A string date in the following format: \%Y-\%m-\%d}
}
\value{
The function returns a dataframe containing all the events saved by BLOOM
}
\description{
Get records in BLOOM fleet register
}
