% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fishing_gears.R
\name{get_fishing_gears}
\alias{get_fishing_gears}
\title{Get MMSI information for all vessels}
\usage{
get_fishing_gears()
}
\value{
The function returns a dataframe containing the MMSI with dates
}
\description{
Get MMSI information for all vessels
}
