% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fleet_repartition.R
\name{get_fleet_repartition}
\alias{get_fleet_repartition}
\title{Get fleet repartition}
\usage{
get_fleet_repartition(
  country_iso3 = NULL,
  start_date = "2024-01-01",
  end_date = Sys.Date(),
  size_breaks = c(0, 12, 18, 24, 40, Inf),
  categorized = TRUE
)
}
\arguments{
\item{country_iso3}{A string ISO3 code. One of "BEL", "BGR", "HRV", "CYP", "DNK", "EST", "FIN", "FRA", DEU", "GRC", "GBR", "IRL", "ITA", "LVA", "LTU", "MLT", "NLD", "POL", "PRT", "ROU", "SVN", "ESP", "SWE"}

\item{start_date}{A string date in the following format: \%Y-\%m-\%d}

\item{end_date}{A string date in the following format: \%Y-\%m-\%d}

\item{size_breaks}{Vessel length breaks array (LOA)}

\item{categorized}{logical. If TRUE then gear are categorized (level 1 with subcategory when available) following the FAO ISSCFG standards (2016). If FALSE, level 2 is used.}
}
\value{
Returns a size dataframe
}
\description{
The classification is based solely on the main gear (order = 0).

If a vessel shows multiple lengths (i.e. main gear) in the given period, the length (i.e. main gear) covering the longer period of time is used.

The gear categorization follows the FAO ISSCFG standards (2016)
\url{https://openknowledge.fao.org/server/api/core/bitstreams/830259c5-cbba-49f8-ae0d-819cd54356d3/content}.
}
\examples{
get_fleet_repartition('FRA', '2024-01-01', '2024-12-31', c(0, 12, 18, 24, 40, Inf))

}
