% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imo.R
\name{get_imo}
\alias{get_imo}
\title{Get IMO/UVI for all vessels}
\usage{
get_imo()
}
\value{
The function returns a dataframe containing all the IMO/UVI registered associated to CFR ids (found in CFR through \code{get_cfr}).
}
\description{
Get IMO/UVI for all vessels
}
