% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mmsi.R
\name{get_mmsi}
\alias{get_mmsi}
\title{Get MMSI information for all vessels}
\usage{
get_mmsi()
}
\value{
The function returns a dataframe containing the MMSI with dates
}
\description{
Get MMSI information for all vessels
}
