% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vessel.R
\name{get_vessel}
\alias{get_vessel}
\title{Get information associated to a vessel ID}
\usage{
get_vessel(id = "FRA000669307", type = "cfr")
}
\arguments{
\item{id}{A vessel identifier (CFR, MMSI, UVI, IRCS, external marking or registration number)}

\item{type}{identifier type (cfr, mmsi, uvi, ircs, external_marking, registration_number)}
}
\value{
The function returns a list containing dataframes with the historical information associated to a vessel ID
}
\description{
Get information associated to a vessel ID
}
