% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fishing_gears.R
\name{get_fishing_gears}
\alias{get_fishing_gears}
\title{Get main fishing gears for all vessels}
\usage{
get_fishing_gears()
}
\value{
The function returns a dataframe containing main fishing gear (order = 0) records associated to CFR ids (found in CFR through \code{get_cfr})
}
\description{
Get main fishing gears for all vessels
}
