% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mmsi.R
\name{get_mmsi}
\alias{get_mmsi}
\title{Get MMSI for all vessels}
\usage{
get_mmsi()
}
\value{
The function returns a dataframe containing all the MMSI registered associated to CFR ids (found in CFR through \code{get_cfr}).
}
\description{
Get MMSI for all vessels
}
